client: Tidy up visleaf threading.
To be fair this is a bit ugly since we extract `RenderableInfo_t` and make it publicly accessible, however
we do take advantage of this to squeeze in a bool value to track if we need to call `ComputeFxBlend()`, saving
us from the additional CUtlVector and having to copy over the values every. single. time.

v2: Minor tweaks for NostalgiaFortress - babaisyou